refactor styles into a separate .scss file
did this so regular and responsive styles can all be in one place
