+++ apps/src/templates/marketingSkinnyBanner.scss
@@ -1,0 +1,70 @@
+// Variables
+$font-regular: 'Gotham 4r', sans-serif;
+$font-bold: 'Gotham 7r', sans-serif;
+$width-md: 970px;
+$width-sm: 600px;
+
+// Styles
+a:hover {
+  text-decoration: none;
+}
+
+aside {
+  box-sizing: content-box;
+  height: 120px;
+  overflow: auto;
+  margin: 4em auto 5em;
+  padding: 0 2.5em;
+  background: url("/blockly/media/csa-skinny-banner-bg.svg") center no-repeat;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  flex-wrap: wrap;
+
+  @media screen and (max-width: $width-sm) {
+    height: auto;
+    margin-top: 2em;
+    padding: 1.5em 1.15em;
+    background: linear-gradient(180deg, #51466E 0%, #75659F 100%);
+    justify-content: center;
+    text-align: center;
+  }
+
+  .text-wrapper {
+    width: 370px;
+
+    h1 {
+      color: white;
+      font-family: $font-bold;
+      font-size: 1.65em;
+      margin: 0;
+    }
+
+    p {
+      color: white;
+      font-family: $font-regular;
+      font-size: 1.15em;
+    }
+  }
+
+  img {
+    width: 315px;
+    margin-bottom: -8px;
+
+    @media screen and (max-width: $width-md) {
+      display: none;
+    }
+  }
+
+  span {
+    font-family: $font-bold;
+    color: white;
+    background: #FFA400;
+    padding: 0.65em 2em;
+    border-radius: 4px;
+
+    @media screen and (max-width: $width-sm) {
+      margin: 1em auto;
+    }
+  }
+}
